pyproject.toml
[project]
name = "aether"
version = '1.0'
requires-python = ">=3.12"
dependencies = [
    "lightning>=2.6.0",
    "torch>=2.9.1",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper==1.2.0",
    "torchvision>=0.24.1",
    "torchaudio>=2.9.1",
    "rootutils>=1.0.7",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "geoclip>=1.2.0",
    "wandb>=0.23.1",
    "rasterio>=1.4.3",
    "pytest>=9.0.2",
    "geopandas>=1.1.1",
    "xarray>=2025.12.0",
    "rioxarray>=0.20.0",
    "utm>=0.8.1",
    "geopy>=2.4.1",
    "tqdm>=4.67.1",
    "sh>=2.2.2",
    "pytest-cov>=7.0.0",
]

[project.optional-dependencies]
create-data = [
    "geemap>=0.36.6",
    "pipreqs>=0.5.0",
]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]
